requirements.txt
blinker
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click
colorama
cryptography==46.0.7
Flask
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-dialogflow==2.47.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
idna==3.11
itsdangerous
Jinja2
MarkupSafe
packaging
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
requests==2.33.1
typing_extensions==4.15.0
urllib3==2.6.3
waitress
Werkzeug
